Separate TableProvider to separate module
